+++ blade-freemarker/src/main/java/com/blade/demo/freemarker/FreeMarkerConfig.java
@@ -1,8 +1,9 @@
-package tpl;
+package com.blade.demo.freemarker;
 
 import com.blade.Blade;
 import com.blade.event.BeanProcessor;
 import com.blade.ioc.annotation.Bean;
+import com.blade.mvc.view.template.FreeMarkerTemplateEngine;
 
 /**
  * Created by Administrator on 2017/12/24.
@@ -11,6 +12,6 @@ import com.blade.ioc.annotation.Bean;
 public class FreeMarkerConfig  implements BeanProcessor{
     @Override
     public void processor(Blade blade) {
-        blade.templateEngine(new FreeMarkerTemplate());
+        blade.templateEngine(new FreeMarkerTemplateEngine());
     }
 }
